aussprachetrainer/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.3
azure-cognitiveservices-speech==1.31.0
azure-common==1.1.28
azure-core==1.29.3
azure-identity==1.14.0
azure-keyvault-secrets==4.7.0
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.1.0
cryptography==41.0.4
cycler==0.11.0
distlib==0.3.7
Django==4.1.10
filelock==3.12.2
fonttools==4.40.0
fontawesomefree==6.4.2
idna==3.4
isodate==0.6.1
joblib==1.3.1
kiwisolver==1.4.4
kombu==5.3.1
matplotlib==3.7.2
meteostat==1.6.5
msal==1.23.0
msal-extensions==1.0.0
numpy
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
platformdirs==3.10.0
portalocker==2.7.0
prompt-toolkit==3.0.39
pycparser==2.21
pydub==0.25.1
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3
redis==5.0.0
requests==2.31.0
scikit-learn==1.3.0
scipy
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.1.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
virtualenv==20.24.3
wcwidth==0.2.6
sentry_sdk
gunicorn==21.2.0
django-cors-headers==4.0.0
psycopg2-binary==2.9.9
django-celery-beat==2.5.0 